Extract the seeding of user account in another service.
